Hello prints "Hello Gophers!" to the terminal
Hello prints "Hello Gophers!" and draws an ASCII art Gopher
to the terminal.

The source code is documented to describe how the program
works and to outline the major parts of a go program.

The program is intended as a test program to test that a
go installation is correct.